Pass Rack env to `map_attributes` and `subject`
Having access to the Rack env in the app-specific logic of the receiver
allows for attributes or other information to be recorded prior to
the authentication event, and then used while building the Subject.

First use case for this is using an invite code to resolve a Subject
which doesn't yet have its persistent identifier set.