Add TestAuthenticator to assist Capybara examples
Login via the existing mechanism is quite messy in Capybara examples,
because the redirect to Rapid Connect can't be trapped and stubbed out.

This provides a hook and an alternate code path so that test code can
inject a JWT which will be presented with a simple 'Login' button to
click in test cases.